RDCC-958: moved ErrorConstants to 'constants' package from 'advice' package
